[AD RACING][Pay ment]

+++ ad_racing_be/src/main/java/com/example/ad_racing_be/order/dto/IOderDto.java
@@ -2,5 +2,5 @@ package com.example.ad_racing_be.order.dto;
 
 public interface IOderDto {
     String getDateTime();
-    Double getPrice();
+    Float getPrice();
 }
